unify HiveFunctionDefinitionFactory with HiveTableFunctory and fix unit tests
