+++ flink-table/flink-table-common/src/main/java/org/apache/flink/table/catalog/Catalog.java
@@ -34,7 +34,6 @@ import org.apache.flink.table.catalog.exceptions.TableNotPartitionedException;
 import org.apache.flink.table.catalog.exceptions.TablePartitionedException;
 import org.apache.flink.table.catalog.stats.CatalogColumnStatistics;
 import org.apache.flink.table.catalog.stats.CatalogTableStatistics;
-import org.apache.flink.table.factories.FunctionDefinitionFactory;
 import org.apache.flink.table.factories.TableFactory;
 
 import java.util.List;
@@ -48,22 +47,12 @@ import java.util.Optional;
 public interface Catalog {
 
 	/**
-	 * Get an optional {@link TableFactory} instance that's responsible for generating source/sink for tables
-	 * stored in this catalog.
+	 * Get an optional {@link TableFactory} instance that's responsible for generating table-related
+	 * instances stored in this catalog, instances such as source/sink and function definitions.
 	 *
 	 * @return an optional TableFactory instance
 	 */
 	default Optional<TableFactory> getTableFactory() {
-		return Optional.empty();
-	}
-
-	/**
-	 * Get an optional {@link FunctionDefinitionFactory} instance that's responsible for generating function definitions
-	 * for catalog functions in this catalog.
-	 *
-	 * @return an optional FunctionDefinitionFactory instance
-	 */
-	default Optional<FunctionDefinitionFactory> getFunctionDefinitionFactory() {
 		return Optional.empty();
 	}
